# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Symbiota API",
        "contact": {
            "name": "Symbiota Support Hub Team",
            "email": "symbiota@asu.edu"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "../"
        }
    ],
    "paths": {
...
        "/api/v2/occurrence/annotation/search": {
            "get": {
                "tags": [
                    ""
                ],
                "operationId": "/api/v2/occurrence/annotation/search",
                "parameters": [
                    {
                        "name": "collid",
                        "in": "query",
                        "description": "Internal identifier (PK) for collection",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Annoration type (internal, external) ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "internal",
                            "enum": [
                                "internal",
                                "external"
                            ]
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "External source of Annoration (e.g. geolocate) ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fieldName",
                        "in": "query",
                        "description": "Name of occurrence field that was annotated (e.g. recordedBy, eventDate) ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "The start date of a date range the annotation was created (e.g. 2022-02-05) ",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "The end date of a date range the annotation was created (e.g. 2022-02-05) ",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the offset for the search results. A limit of 200 and offset of 100, will get the third page of 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of occurrence edits",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
...
    },
...
}